requirements.txt
beautifulsoup4>=4.12.0
pyyaml>=6.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.1
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.6.0
pdfminer.six>=20220506
psutil>=5.9.0
pynvml>=11.4.1
nvidia-ml-py>=12.535.108
huggingface-hub>=0.23.0
llama-cpp-python>=0.2.0
